some refactoring and utils created
