+++ src/main/kotlin/com/github/etkachev/nxwebstorm/utils/FindSchematics.kt
@@ -3,34 +3,27 @@ package com.github.etkachev.nxwebstorm.utils
 import com.github.etkachev.nxwebstorm.models.SchematicInfo
 import com.github.etkachev.nxwebstorm.ui.settings.PluginSettingsState
 import com.google.gson.JsonObject
+import com.google.gson.JsonParser
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.roots.ProjectRootManager
 import com.intellij.openapi.vfs.VirtualFile
-import com.intellij.psi.PsiManager
+import com.intellij.psi.PsiFile
 import com.intellij.psi.search.FilenameIndex
 import com.intellij.psi.search.GlobalSearchScopes
 
 data class CollectionInfo(val json: JsonObject, val file: VirtualFile)
-data class PackageJsonInfo(val packageName: String, val json: JsonObject, val file: VirtualFile)
 class FindSchematics(private val project: Project, private val externalLibs: Array<String>) {
   var jsonFileReader = ReadFile(project)
+  var packageJsonHelper = PackageJsonHelper(project)
   var schematicPropName = "schematics"
   var nodeModulesFolder = "node_modules"
 
-  private fun getPackageFileInfo(directory: String): PackageJsonInfo? {
-    val packageFile = jsonFileReader.findVirtualFile("$nodeModulesFolder/$directory/package.json") ?: return null
-    return getPackageFileByVirtualFile(packageFile)
+  fun findByExternalLibs(): Map<String, SchematicInfo> {
+    val schematicMaps = externalLibs.mapNotNull { dir -> getSchematicsFromNodeModulesDirectory(dir) }
+    return foldListOfMaps(schematicMaps.toTypedArray())
   }
 
-  private fun getPackageFileByVirtualFile(file: VirtualFile): PackageJsonInfo? {
-    val packageFileJson = jsonFileReader.readJsonFromFile(file) ?: return null
-    val packageName = (if (packageFileJson.has("name")) packageFileJson["name"].asString else null) ?: return null
-    return PackageJsonInfo(packageName, packageFileJson, file)
-  }
-
-  private fun findAllPackageJsonFiles(): Map<String, SchematicInfo> {
-    val root = ProjectRootManager.getInstance(project).contentRoots[0]
-    val psiDir = PsiManager.getInstance(project).findDirectory(root) ?: return emptyMap()
+  fun scanAllForExternalSchematics(): Map<String, SchematicInfo> {
+    val psiDir = getRootPsiDirectory(project) ?: return emptyMap()
 
     val nodeModules = psiDir.findSubdirectory(nodeModulesFolder) ?: return emptyMap()
     val splitProjectRootPath = psiDir.virtualFile.path.split("/")
@@ -40,31 +33,56 @@ class FindSchematics(private val project: Project, private val externalLibs: Arr
       "package.json",
       GlobalSearchScopes.directoriesScope(project, true, nodeModules.virtualFile)
     ).mapNotNull { f ->
-      val (packageName, packageFileJson, packageFile) = getPackageFileByVirtualFile(f.virtualFile)
+      val (packageName, packageFileJson, packageFile) = packageJsonHelper.getPackageFileByVirtualFile(f.virtualFile)
         ?: return@mapNotNull null
       val (schematicsOptions, schematicCollection) = getSchematicOptions(packageFileJson, packageFile)
         ?: return@mapNotNull null
       val splitFullFileLocation = f.parent!!.virtualFile.path.split("$projectRootFolder/$nodeModulesFolder/")
       val nodeModulesFileLocation = splitFullFileLocation[1]
       val packageFileLocation = "$nodeModulesFolder/$nodeModulesFileLocation"
-      val schematicEntries =
-        schematicsOptions.entrySet().toTypedArray().fold(mutableMapOf<String, SchematicInfo>(), { acc, e ->
-          val value = e.value.asJsonObject
-          if (value.has("hidden") && value["hidden"].asBoolean) {
-            return@fold acc
-          }
-
-          val relativePath = getRelativePath(nodeModulesFileLocation, value, schematicCollection) ?: return@fold acc
-          val schematicsFileLocation = "$packageFileLocation$relativePath"
-          val id = generateUniqueSchematicKey(packageName, e.key)
-          val description = if (value.has("description")) value["description"].asString else null
-          acc[id] = SchematicInfo(schematicsFileLocation, description)
-          return@fold acc
-        })
-      val results = schematicEntries.toMap()
+      val results = getSchematicEntries(
+        schematicsOptions,
+        nodeModulesFileLocation,
+        schematicCollection,
+        packageName,
+        fileLocationMapper = { path -> "$packageFileLocation$path" })
       return@mapNotNull if (results.isNotEmpty()) results else null
     }.toTypedArray()
-    return foldListOfSchematicMaps(packageJsonFiles)
+    return foldListOfMaps(packageJsonFiles)
+  }
+
+  private fun getSchematicEntries(
+    schematicOptions: JsonObject,
+    directory: String,
+    schematicCollection: VirtualFile,
+    packageName: String,
+    fileLocationMapper: (path: String) -> String
+  ): Map<String, SchematicInfo> {
+    return schematicOptions.entrySet().toTypedArray().fold(mutableMapOf<String, SchematicInfo>(), { acc, e ->
+      val value = e.value.asJsonObject
+      if (value.has("hidden") && value["hidden"].asBoolean) {
+        return@fold acc
+      }
+
+      val relativePath = getRelativePath(directory, value, schematicCollection) ?: return@fold acc
+      val fileLocation = fileLocationMapper(relativePath)
+      val id = generateUniqueSchematicKey(packageName, e.key)
+      val description = if (value.has("description")) value["description"].asString else null
+      acc[id] = SchematicInfo(fileLocation, description)
+      return@fold acc
+    }).toMap()
+  }
+
+  private fun getSchematicsFromNodeModulesDirectory(directory: String): Map<String, SchematicInfo>? {
+    val packageJson = packageJsonHelper.getPackageFileInfo("$nodeModulesFolder/$directory") ?: return null
+    val (packageName, packageFileJson, packageFile) = packageJson
+    val (schematicsOptions, schematicCollection) = getSchematicOptions(packageFileJson, packageFile) ?: return null
+    return getSchematicEntries(
+      schematicsOptions,
+      directory,
+      schematicCollection,
+      packageName,
+      fileLocationMapper = { path -> "$nodeModulesFolder/$directory$path" })
   }
 
   private fun getSchematicOptions(packageFileJson: JsonObject, packageFile: VirtualFile): CollectionInfo? {
@@ -86,54 +104,27 @@ class FindSchematics(private val project: Project, private val externalLibs: Arr
     val splitFile = schemaFile.path.split(directory)
     return if (splitFile.count() == 2) splitFile[1] else null ?: return null
   }
-
-  private fun getSchematicsFromDirectory(directory: String): Map<String, SchematicInfo>? {
-    val packageJson = getPackageFileInfo(directory) ?: return null
-    val (packageName, packageFileJson, packageFile) = packageJson
-    val (schematicsOptions, schematicCollection) = getSchematicOptions(packageFileJson, packageFile) ?: return null
-    val schematicEntries =
-      schematicsOptions.entrySet().toTypedArray().fold(mutableMapOf<String, SchematicInfo>(), { acc, e ->
-        val value = e.value.asJsonObject
-        if (value.has("hidden") && value["hidden"].asBoolean) {
-          return@fold acc
-        }
-        val relativePath = getRelativePath(directory, value, schematicCollection) ?: return@fold acc
-        val fileLocation = "$nodeModulesFolder/$directory$relativePath"
-        val id = generateUniqueSchematicKey(packageName, e.key)
-        val description = if (value.has("description")) value["description"].asString else null
-        acc[id] = SchematicInfo(fileLocation, description)
-        return@fold acc
-      })
-
-    return schematicEntries.toMap()
-  }
-
-  fun findByExternalLibs(): Map<String, SchematicInfo> {
-    val schematicMaps = externalLibs.mapNotNull { dir -> getSchematicsFromDirectory(dir) }
-    return foldListOfSchematicMaps(schematicMaps.toTypedArray())
-  }
-
-  private fun foldListOfSchematicMaps(schematicMaps: Array<Map<String, SchematicInfo>>): Map<String, SchematicInfo> {
-    return schematicMaps.fold(mutableMapOf<String, SchematicInfo>(), { acc, e ->
-      for (key in e.keys) {
-        val info = e[key] ?: continue
-        acc[key] = info
-      }
-      return@fold acc
-    }).toMap()
-  }
-
-  fun scanAllForExternalSchematics(): Map<String, SchematicInfo> {
-    return findAllPackageJsonFiles()
-  }
 }
 
 class FindAllSchematics(private val project: Project) {
-  fun findAll(): Map<String, SchematicInfo> {
-    val customSchematics = GetNxData(project).getCustomSchematics()
+  var defaultToolsSchematicDir = "/tools/schematics"
+  var configToolsSchematicDir: String
+  private val cleanedUpConfigToolsSchematicDir: String
+    get() = if (configToolsSchematicDir.startsWith("/")) configToolsSchematicDir else "/$configToolsSchematicDir"
+  private val toolsSchematicDir: String
+    get() = if (configToolsSchematicDir.isBlank()) defaultToolsSchematicDir else cleanedUpConfigToolsSchematicDir
+  private val splitSchematicDir: Array<String>
+    get() = toolsSchematicDir.split("/").mapNotNull { s -> if (s.isBlank()) null else s }.toTypedArray()
+
+  init {
     val settings: PluginSettingsState = PluginSettingsState.instance
-    val findExplicitSchematics = settings.scanExplicitLibs
-    if (findExplicitSchematics) {
+    configToolsSchematicDir = settings.customSchematicLocation
+  }
+
+  fun findAll(): Map<String, SchematicInfo> {
+    val customSchematics = getCustomSchematics()
+    val settings: PluginSettingsState = PluginSettingsState.instance
+    if (settings.scanExplicitLibs) {
       val others = settings.externalLibs.split(",").mapNotNull { value ->
         val trimmed = value.trim()
         val final = if (trimmed.isEmpty()) null else trimmed
@@ -149,5 +140,27 @@ class FindAllSchematics(private val project: Project) {
       val allScanned = FindSchematics(project, emptyArray()).scanAllForExternalSchematics()
       return flattenMultipleMaps(customSchematics, allScanned)
     }
+  }
+
+  private fun getCustomSchematics(): Map<String, SchematicInfo> {
+    val rootPsiDirectory = getRootPsiDirectory(project)
+    val schematics = findPsiDirectoryBySplitFolders(splitSchematicDir, rootPsiDirectory) ?: return emptyMap()
+    val files = FilenameIndex.getFilesByName(
+      project, "schema.json",
+      GlobalSearchScopes.directoriesScope(project, true, schematics.virtualFile)
+    )
+    return files.mapNotNull { file -> getIdsFromSchema(file) }.toMap()
+  }
+
+  private fun getIdsFromSchema(file: PsiFile): Pair<String, SchematicInfo>? {
+    val json = JsonParser.parseString(file.text).asJsonObject ?: return null
+    if (!json.has("id")) {
+      return null
+    }
+    val id = json.get("id").asString
+    val fileLocation = "$toolsSchematicDir/$id/schema.json"
+    val info = SchematicInfo(fileLocation)
+    val uniqueId = generateUniqueSchematicKey("workspace-schematic", id)
+    return uniqueId to info
   }
 }
